move http driver tests
