+++ packages/drivers/http/package.json
@@ -5,6 +5,7 @@
   "author": "Fabien JUIF <fabien.juif@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "k-ramel": "0.7.0",
     "cross-env": "5.1.3",
     "eslint": "4.17.0"
   }
