+++ packages/drivers/http/src/index.spec.js
@@ -1,0 +1,5 @@
+/* eslint-env jest */
+import driver from './index'
+import tests from '../misc/test-suite'
+
+tests(driver)
